Add endpoint for file upload
This commit extends the `/files/{file}` endpoint to POST requests which
create new files from the request body as contents.

Note that currently the endpoint reads the full request body into
memory which is sub-optimal. It'd might be possible to improve upon this
in the future (e.g., by copying the body directly into an open file
while still reading it).